+++ danilenko.ekaterina/I2/MaxElements.h
@@ -1,6 +1,6 @@
 #ifndef MAXELEMENTS
 #define MAXELEMENTS
 
-size_t CountMaxElements(const int *array, size_t size);
+int CountMaxElements(const int *array, size_t size);
 
 #endif
