+++ danilenko.ekaterina/I2/PositiveRepeatingNumber.h
@@ -1,6 +1,6 @@
 #ifndef POSITIVEREPEATINGNUMBER
 #define POSITIVEREPEATINGNUMBER
 
-size_t PosRepeatingNum(const int* array, size_t size);
+int PosRepeatingNum(const int* array, size_t size);
 
 #endif
